HoldFps - change from int to double
04.07.26	- GetAdapterInfo - change if(!driverdescription || !driverversion)
	  to if(!driverdescription[0] || !driverversion[0])
	  Remove unused Devices variable
15.07.26	- HoldFps - change from int to double

+++ SPOUTSDK/SpoutGL/Spout.h
@@ -161,7 +161,7 @@ class SPOUT_DLLEXP Spout : public spoutGL {
 	// Return frame count status
 	bool IsFrameCountEnabled();
 	// Frame rate control
-	void HoldFps(int fps);
+	void HoldFps(double fps);
 	// Signal sync event 
 	void SetFrameSync(const char* name = nullptr);
 	// Wait or test for a sync event
